darkguard-openenv/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "darkguard-openenv"
version = "0.1.0"
description = "DarkGuard OpenEnv environment space"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.9.0",
  "requests>=2.32.0",
  "httpx>=0.27.0",
  "openenv-core[core]>=0.2.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
